InsureTM/requirements.txt
# InsureTM — Backend Dependencies
# Generated from the project virtual environment

# Django core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
daphne==4.1.0
channels==4.1.0

# Database
psycopg2-binary==2.9.11

# Production server
gunicorn==21.2.0

# File handling
Pillow==12.1.1

# AI / Analytics
groq>=0.13.0
httpx>=0.27.0,<0.29.0
vanna==2.0.2
scikit-learn==1.8.0
numpy==2.4.2
pandas==3.0.0
joblib==1.3.2
SQLAlchemy==2.0.46
plotly==6.5.2

# HTTP
requests==2.32.5

# PDF Generation
fpdf2==2.8.2

# Swagger / OpenAPI
drf-spectacular==0.27.1
